0924　ログイン機能とか作ったよ

+++ index.js
@@ -1,14 +1,18 @@
 const express = require("express")
 const app = express()
-const port = 5050;
+const port = process.env.PORT || 5050;
+const cors = require("cors")
+app.use(cors())
 app.use(express.urlencoded({ extended:true }))
 app.use(express.json())
+const jwt = require("jsonwebtoken")
+const auth = require("./utils/auth")
 const connectDB = require("./utils/database")
-const { ItemModel } = require("./utils/schemaModels")
+const { ItemModel, UserModel } = require("./utils/schemaModels")
 
 //ITEM functions
 //Create Item
-app.post("/item/create", async(req, res) => {
+app.post("/item/create", auth, async(req, res) => {
   try {
   connectDB()
   await ItemModel.create(req.body)
@@ -41,22 +45,32 @@ app.get("/item/:id", async(req, res) => {
 })
 
 //Update Item
-app.put("/item/update/:id", async(req, res) => {
+app.put("/item/update/:id", auth, async(req, res) => {
   try{
     await connectDB()
-    const singleItem = await ItemModel.updateOne({_id: req.params.id}, req.body)
-    return res.status(200).json({message: "アイテム編集成功（シングル）", singleItem: singleItem})
+    const singleItem = await ItemModel.findById(req.params.id)
+    if (singleItem.email === req.body.email){
+      await ItemModel.updateOne({_id: req.params.id}, req.body)
+      return res.status(200).json({message: "アイテム編集成功（シングル）", singleItem: singleItem})
+    }else{
+      throw new Error()
+    }
   }catch(err){
     return res.status(400).json({message: "アイテム編集失敗"})
   }
 })
 
 //Delete Item
-app.delete("/item/delete/:id", async(req, res) => {
+app.delete("/item/delete/:id", auth, async(req, res) => {
   try{
     await connectDB()
-    await ItemModel.deleteOne({_id: req.params.id})
-    return res.status(200).json({message: "アイテム削除成功"})
+    const singleItem = await ItemModel.findById(req.params.id)
+    if(singleItem.email === req.body.email){
+      await ItemModel.deleteOne({_id: req.params.id})
+      return res.status(200).json({message: "アイテム削除成功"})
+    }else{
+      throw new Error()
+    }
   }catch(err){
     return res.status(400).json({message: "アイテム削除失敗"})
   }
@@ -64,8 +78,42 @@ app.delete("/item/delete/:id", async(req, res) => {
 
 //USER functions
 //Register User
+app.post("/user/register", async(req, res) => {
+  try{
+    await connectDB()
+    await UserModel.create(req.body)
+    return res.status(200).json({message: "ユーザー登録成功"})
+  }catch(err){
+    return res.status(400).json({message: "ユーザー登録失敗"})
+  }
+})
+
 //Login User
+const secret_key = "mern-market"
+
+app.post("/user/login", async(req, res) => {
+  try{
+    await connectDB()
+    const savedUserData = await UserModel.findOne({email: req.body.email})
+    if(savedUserData){
+      if(req.body.password === savedUserData.password){
+        const payload = {
+          email: req.body.email,
+        }
+        const token = jwt.sign(payload, secret_key, {expiresIn: "23h"})
+        console.log(token)
+        return res.status(200).json({message: "ログイン成功"})
+      }else{
+        return req.status(400).json({message: "ログイン失敗: パスワードが間違っています"})
+      }
+    }else{
+      return res.status(400).json({message: "ログイン失敗: ユーザー登録をしてください"})
+    }
+  }catch(err){
+    return res.status(400).json({message: "ログイン失敗"})
+  }
+})
 
 app.listen(port, () => {
-  console.log (`Listening onLocalhost port ${port}`)
+  console.log (`Listening on localhost port ${port}`)
 })
